+++ factcheck/cmd/api/internal/handler/topics_it_test.go
@@ -19,7 +19,7 @@ import (
 )
 
 func TestHandlerTopic_Stateful(t *testing.T) {
-	app, cleanup, err := di.InitializeContainerTestV2()
+	app, cleanup, err := di.InitializeContainerTest()
 	if err != nil {
 		panic(err)
 	}
@@ -316,7 +316,7 @@ func TestHandlerTopic_Stateful(t *testing.T) {
 }
 
 func TestHandlerTopic_CountTopicsHome(t *testing.T) {
-	app, cleanup, err := di.InitializeContainerTestV2()
+	app, cleanup, err := di.InitializeContainerTest()
 	if err != nil {
 		panic(err)
 	}
@@ -675,7 +675,7 @@ func TestHandlerTopic_CountTopicsHome(t *testing.T) {
 }
 
 func TestHandlerTopic_ListTopicsHome(t *testing.T) {
-	app, cleanup, err := di.InitializeContainerTestV2()
+	app, cleanup, err := di.InitializeContainerTest()
 	if err != nil {
 		panic(err)
 	}
